Fix fractional BPM not getting sent to VSTs

+++ include/VstSyncController.h
@@ -42,7 +42,7 @@ public:
 
 	void setAbsolutePosition(double ticks);
 	void setPlaybackState(bool enabled);
-	void setTempo(int newTempo);
+	void setTempo(double newTempo);
 	void setTimeSignature(int num, int denom);
 	void startCycle(int startTick, int endTick);
 	void stopCycle();
