+++ src/core/VstSyncController.cpp
@@ -82,7 +82,7 @@ void VstSyncController::setPlaybackState(bool enabled)
 
 
 
-void VstSyncController::setTempo(int newTempo)
+void VstSyncController::setTempo(double newTempo)
 {
 	if (!m_syncData) { return; }
